Rename org.bdgenomics.adam.rdd.features package to o.b.a.rdd.feature
