+++ adam-core/src/main/scala/org/bdgenomics/adam/rdd/ADAMContext.scala
@@ -47,7 +47,7 @@ import org.bdgenomics.adam.projections.{
   Projection
 }
 import org.bdgenomics.adam.rdd.contig.NucleotideContigFragmentRDD
-import org.bdgenomics.adam.rdd.features._
+import org.bdgenomics.adam.rdd.feature._
 import org.bdgenomics.adam.rdd.fragment.FragmentRDD
 import org.bdgenomics.adam.rdd.read.{ AlignedReadRDD, AlignmentRecordRDD, UnalignedReadRDD }
 import org.bdgenomics.adam.rdd.variation._
